refactor: add error package
