+++ http-proto/httperr/errors.go
@@ -1,0 +1,7 @@
+package httperr
+
+import "errors"
+
+var (
+	ErrIncompleteHeader = errors.New("incomplete headers")
+)
